`security-audit` refactor/improvements: 1) add new command `security-audit` w/ sub-commands (a) `report` (replaces `security-audit-report` cmd), and (b) `sync` -- uses KA endpoint clear_security_data to clear/reset/recalculate vault security scores admin-side
